Updated devices to no longer use device_proxy. Fixed some signal types

+++ desy_bluesky/devices/omsvme58.py
@@ -25,12 +25,12 @@ from .fsec_readable_device import FSECReadableDevice
 
 class OmsVME58Motor(FSECReadableDevice, Movable, Stoppable):
     Position: A[SignalRW[float], Format.HINTED_UNCACHED_SIGNAL]
-    SlewRate: A[SignalRW[float], Format.CONFIG_SIGNAL]
-    SlewRateMax: A[SignalRW[float], Format.CONFIG_SIGNAL]
+    SlewRate: A[SignalRW[int], Format.CONFIG_SIGNAL]
+    SlewRateMax: A[SignalRW[int], Format.CONFIG_SIGNAL]
     Conversion: A[SignalRW[float], Format.CONFIG_SIGNAL]
-    Acceleration: A[SignalRW[float], Format.CONFIG_SIGNAL]
+    Acceleration: A[SignalRW[int], Format.CONFIG_SIGNAL]
     StopMove: SignalX
-    StepPositionController: SignalRW[float]
+    StepPositionController: SignalRW[int]
     PositionEncoder: SignalR[float]
     PositionEncoderRaw: SignalR[float]
